Small config changes (#845)
* Adding /etc/cowrie/cowrie.cfg to possible configurations

We want also to look for /etc/cowrie/cowrie.cfg as a possible
configuration.

* Write ssh host keys into /var/lib/cowrie

Dynamical data should be written into /var/lib/cowrie.
I know that OpenSSH is doing this but we are not OpenSSH and we should
have only stuff written below /var.
So /var/log/cowrie and /var/lib/cowrie.

* Moving log and dl below var/, cleanup old folders

This helps keeping the writes of the daemon in one place and makes it
easier later when building a cowrie package.

Old paths have been removed from the repository to keep it clean.

* Fixing wrong log path

Path should be var/log/cowrie not var/log

* Fixing json output

The json output was not configured to use the default log path. This has
been fixed now.

* Change order of config reads

According to @micheloosterhof there is an order of precedence here.